Move scalar mul things we want to upstream to jubjub to their own crate
