--- src/traits.rs
@@ -1,48 +1,0 @@
-use std::borrow::Borrow;
-
-use crate::Scalar;
-
-pub trait NonAdjacentForm {
-    fn non_adjacent_form(&self, w: usize) -> [i8; 256];
-}
-
-/// A trait for variable-time multiscalar multiplication without precomputation.
-pub trait VartimeMultiscalarMul {
-    /// The type of point being multiplied, e.g., `AffinePoint`.
-    type Point;
-
-    /// Given an iterator of public scalars and an iterator of
-    /// `Option`s of points, compute either `Some(Q)`, where
-    /// $$
-    /// Q = c\_1 P\_1 + \cdots + c\_n P\_n,
-    /// $$
-    /// if all points were `Some(P_i)`, or else return `None`.
-    fn optional_multiscalar_mul<I, J>(scalars: I, points: J) -> Option<Self::Point>
-    where
-        I: IntoIterator,
-        I::Item: Borrow<Scalar>,
-        J: IntoIterator<Item = Option<Self::Point>>;
-
-    /// Given an iterator of public scalars and an iterator of
-    /// public points, compute
-    /// $$
-    /// Q = c\_1 P\_1 + \cdots + c\_n P\_n,
-    /// $$
-    /// using variable-time operations.
-    ///
-    /// It is an error to call this function with two iterators of different lengths.
-    fn vartime_multiscalar_mul<I, J>(scalars: I, points: J) -> Self::Point
-    where
-        I: IntoIterator,
-        I::Item: Borrow<Scalar>,
-        J: IntoIterator,
-        J::Item: Borrow<Self::Point>,
-        Self::Point: Clone,
-    {
-        Self::optional_multiscalar_mul(
-            scalars,
-            points.into_iter().map(|p| Some(p.borrow().clone())),
-        )
-        .unwrap()
-    }
-}
